Refactor damage-specific pieces out of distribute system, all tests passing
